apply cargo clippy suggestions
apply numerous cargo clippy suggestions. Removing proc_map parameter
from add_process_recursive was fine, because proc_map is only used to
resolve pids to indices in the proc_tree vector while we are building
the proc_tree structure. Once it is built, each node's 'children'
attribute is a list of indices into proc_tree. These are not pids, so
there is no more need to resolve pids to indices, so indeed, proc_map is
not needed at this point.