remove api methods not being used
remove the whole entry api and add_at, which were only being used by the
library crate itself. I don't think either of the two programs I'm
writing will use these. This cuts down on the amount of code in the
library crate. I'm not sure if this change will make the library crate
more or less usable in the long run.