use KB instead of KiB in untitled-smaps-poller
use KB instead of KiB in untitled-smaps-poller. for now, while gnuplot
puts ticks at multiples of a power of 10, this will make graphs more
coherent. I may want to have the ticks be a multiple of a power of two
later but that would take me a little bit of time